+++ pydecor/constants.py
@@ -1,0 +1,5 @@
+"""
+This rather skinny module just contains constants for use elsewhere.
+"""
+
+LOG_CALL_FMT_STR = '{name}(*{args}, **{kwargs}) -> {result}'